Refactor shape drawing API and add draw module
Moved shape drawing functions (circle, line, point, rect, polygon, etc.) from the renderer namespace to a new draw namespace and module. Added Draw.hpp and draw.cpp for implementation, updated Python bindings and type stubs, and revised CMakeLists.txt and _core.cpp to include the new module. Updated pyproject.toml version to 0.1.1. This change improves API clarity and separation of concerns for rendering and shape drawing.
