+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "kraken-engine"
-version = "0.1.0"
+version = "0.1.1"
 authors = [
   { name = "Derrick Martinez" },
 ]
